pyproject.toml
[tool.poetry]
name = "django net"
version = "0.1.0"
description = "Social network"
authors = ["Neolize"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"
pillow = "^9.5.0"
psycopg2 = "^2.9.6"
djoser = "^2.2.0"
python-dotenv = "^1.0.0"
djangorestframework-simplejwt = "^5.2.2"
django-mptt = "^0.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
